Chore: Move files into single api/ functions
