+++ api/rules/index.js
@@ -1,0 +1,6 @@
+const { send } = require('micro');
+const db = require('../../db');
+
+module.exports = (req, res) => send(res, 200, {
+  rules: db.get('taxRules').value()
+});
