+++ lib/index.js
@@ -1,6 +1,6 @@
 const _ = require('lodash');
 const moment = require('moment');
-const db = require('./db');
+const db = require('../db');
 const FormulaParser = require('hot-formula-parser').Parser;
 const parser = new FormulaParser();
 const Currency = require('currency.js');
