+++ tests/integration/arg.test.js
@@ -1,4 +1,4 @@
-const { getTaxes } = require('../../getTax');
+const { getTaxes } = require('../../lib');
 
 describe('Argentina', () => {
   describe('Inscripto', () => {
