+++ tests/integration/ksa.test.js
@@ -1,4 +1,4 @@
-const { getTaxes } = require('../../getTax');
+const { getTaxes } = require('../../lib');
 
 describe('KSA VAT', () => {
   describe('DRUG', () => {
